Move Topology class to topology.py
